Backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.26.0,<2.0
Werkzeug>=3.0.0
easyocr>=1.7.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0

# AI & Text-to-Speech
google-generativeai>=0.8.0
gTTS>=2.4.0
pydub>=0.25.1

# Optional: For better OCR (uncomment if needed)
# pytesseract==0.3.10
# google-cloud-vision==3.5.0

